# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin API 文档",
        "title": "Gin Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "25076778@qq.com"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:8080",
    "paths": {
...
        "/api/v1/system-config": {
            "get": {
                "description": "系统配置列表",
                "tags": [
                    "系统配置管理"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "是否不分页",
                        "name": "notPage",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/request.PageData"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.SystemConfig"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "系统配置创建",
                "tags": [
                    "系统配置管理"
                ],
                "summary": "创建",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "创建参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SystemConfigCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.SystemConfig"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/system-config/{id}": {
            "get": {
                "description": "系统配置详情",
                "tags": [
                    "系统配置管理"
                ],
                "summary": "详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.SystemConfig"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "系统配置更新",
                "tags": [
                    "系统配置管理"
                ],
                "summary": "更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SystemConfigUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/errcode.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "系统配置删除",
                "tags": [
                    "系统配置管理"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/errcode.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errcode.ArgsErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Invalid arguments"
                }
            }
        },
        "errcode.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "errcode.SystemErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
...
        "model.ConfigCategory": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.SystemConfig": {
            "type": "object",
            "properties": {
                "configCategory": {
                    "$ref": "#/definitions/model.ConfigCategory"
                },
                "configCategoryId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "defaultValue": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "optionValue": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "request.PageData": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "数据列表"
                },
                "notPage": {
                    "description": "不分页",
                    "type": "boolean",
                    "example": false
                },
                "page": {
                    "description": "当前页",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
...
        "request.SystemConfigCreate": {
            "type": "object",
            "required": [
                "defaultValue",
                "key",
                "name",
                "optionValue"
            ],
            "properties": {
                "configCategoryId": {
                    "type": "integer"
                },
                "defaultValue": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "optionValue": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "request.SystemConfigUpdate": {
            "type": "object",
            "required": [
                "defaultValue",
                "key",
                "name",
                "optionValue"
            ],
            "properties": {
                "configCategoryId": {
                    "type": "integer"
                },
                "defaultValue": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "optionValue": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
    }
}